+++ edgework/models/__init__.py
@@ -1,1 +1,6 @@
 # Models package
+
+from edgework.models.play import Play
+from edgework.models.play_by_play import PlayByPlay
+
+__all__ = ["Play", "PlayByPlay"]